eye_tracker/tracker/requirements.txt
websockets
websocket-client
PyQt6
dlib
opencv-python
imageio
imageio-ffmpeg
quick-queue